requirements.txt
flask
groq
langchain_community
langchain
faiss-cpu
sentence-transformers
langchain-huggingface
langchain_core
langchain_groq
requests
pyyaml
pandas
PyPDF2